Adds S* pattern to support variable number of milliseconds in the decimal part in dates and times
